Refactoring and polishing
